correction due to 64 bit vs. 32 bit
git-svn-id: file:///usr/local/Singular/svn/trunk@12155 2c84dea3-7e68-4137-9b89-c4e89433aadc

+++ Singular/extra.cc
@@ -1,7 +1,7 @@
 /*****************************************
 *  Computer Algebra System SINGULAR      *
 *****************************************/
-/* $Id: extra.cc,v 1.319 2009-10-02 13:51:19 seelisch Exp $ */
+/* $Id: extra.cc,v 1.320 2009-10-02 16:13:30 seelisch Exp $ */
 /*
 * ABSTRACT: general interface to internals of Singular ("system" command)
 */
@@ -2108,7 +2108,7 @@ static BOOLEAN jjEXTENDED_SYSTEM(leftv res, leftv h)
         if (h == NULL) testWrappers(FALSE);  // wrapper tests, no detailed printout
         else if (h->Typ() == INT_CMD)
         {
-          const int detailedOutput = (const int)h->Data();
+          const int detailedOutput = (const int)(long)h->Data();
           testWrappers(detailedOutput == 0 ? FALSE : TRUE);
           // wrapper tests, with or without detailed printout
         }
@@ -2139,9 +2139,9 @@ static BOOLEAN jjEXTENDED_SYSTEM(leftv res, leftv h)
                  (h->next->next->next->next->next == NULL))
         {
           const matrix m           = (const matrix)h->Data();
-          const int k              = (const int)h->next->Data();
-          const int strategy       = (const int)h->next->next->Data();
-          const int cacheEntries   = (const int)h->next->next->next->Data();
+          const int k              = (const int)(long)h->next->Data();
+          const int strategy       = (const int)(long)h->next->next->Data();
+          const int cacheEntries   = (const int)(long)h->next->next->next->Data();
           const long cacheWeight   = (const long)h->next->next->next->next->Data();
           ideal iii = testAllPolyMinorsAsIdeal(m, k, strategy, cacheEntries, cacheWeight);
             // starts the computation of all k x k minors in the
@@ -2164,14 +2164,14 @@ static BOOLEAN jjEXTENDED_SYSTEM(leftv res, leftv h)
                  (h->next->next->next->next->next->next->next->Typ() == INT_CMD))
         {
           const matrix m           = (const matrix)h->Data();
-          const int k              = (const int)h->next->Data();
-          const int strategies     = (const int)h->next->next->Data();
-          const int cacheEntries   = (const int)h->next->next->next->Data();
+          const int k              = (const int)(long)h->next->Data();
+          const int strategies     = (const int)(long)h->next->next->Data();
+          const int cacheEntries   = (const int)(long)h->next->next->next->Data();
           const long cacheWeight   = (const long)h->next->next->next->next->Data();
-          const int dumpMinors     = (const int)h->next->next->next->next->next->Data();
-          const int dumpResults    = (const int)h->next->next->next->next->next->next->Data();
-          const int dumpComplete   = (const int)h->next->next->next->next->next->next->next->Data();
-          const int dumpConsole    = (const int)h->next->next->next->next->next->next->next->next->Data();
+          const int dumpMinors     = (const int)(long)h->next->next->next->next->next->Data();
+          const int dumpResults    = (const int)(long)h->next->next->next->next->next->next->Data();
+          const int dumpComplete   = (const int)(long)h->next->next->next->next->next->next->next->Data();
+          const int dumpConsole    = (const int)(long)h->next->next->next->next->next->next->next->next->Data();
           testAllPolyMinors(m, k, strategies, cacheEntries, cacheWeight, dumpMinors, dumpResults, dumpComplete, dumpConsole);
             // starts the computation of all k x k minors in the
             // provided matrix m (which is assumed to have polynomial
